creado metodo de busqueda, aun no esta implementado en la tabla pero funciona en consola

+++ src/Controller/Forms/Stock/Cls_Item.java
@@ -3,10 +3,10 @@ package Controller.Forms.Stock;
 public class Cls_Item {
     public String ref;
     public String name;
-    public int pvp;
-    public int neto;
+    public double pvp;
+    public double neto;
     public int units;
-    public Cls_Item(String ref, String name, int pvp, int neto, int units) {
+    public Cls_Item(String ref, String name, double pvp, double neto, int units) {
         this.ref = ref;
         this.name = name;
         this.pvp = pvp;
